[PBCKP-245] more mingw compatibility
- pid_t is 64bit in mingw. Lets cast it to long long in most places on
  I/O.
- int64 should be casted to long long as well
- size_t should be printed as %zu/%zd
- stat functino is imported from pgport, and there's no need to
  redefine it again.
- no need to redeclare `__thread` since gcc work with it well.
- arguments and types in launch agent.

+++ src/catchup.c
@@ -137,8 +137,8 @@ catchup_preflight_checks(PGNodeInfo *source_node_info, PGconn *source_conn,
 		}
 		else if (pid > 1) /* postmaster is up */
 		{
-			elog(ERROR, "Postmaster with pid %u is running in destination directory \"%s\"",
-				pid, dest_pgdata);
+			elog(ERROR, "Postmaster with pid %lld is running in destination directory \"%s\"",
+				(long long)pid, dest_pgdata);
 		}
 	}
 
@@ -160,15 +160,15 @@ catchup_preflight_checks(PGNodeInfo *source_node_info, PGconn *source_conn,
 		source_id = get_system_identifier(FIO_DB_HOST, source_pgdata, false); /* same as instance_config.system_identifier */
 
 		if (source_conn_id != source_id)
-			elog(ERROR, "Database identifiers mismatch: we connected to DB id %lu, but in \"%s\" we found id %lu",
-				source_conn_id, source_pgdata, source_id);
+			elog(ERROR, "Database identifiers mismatch: we connected to DB id %llu, but in \"%s\" we found id %llu",
+				(long long)source_conn_id, source_pgdata, (long long)source_id);
 
 		if (current.backup_mode != BACKUP_MODE_FULL)
 		{
 			dest_id = get_system_identifier(FIO_LOCAL_HOST, dest_pgdata, false);
 			if (source_conn_id != dest_id)
-			elog(ERROR, "Database identifiers mismatch: we connected to DB id %lu, but in \"%s\" we found id %lu",
-				source_conn_id, dest_pgdata, dest_id);
+			elog(ERROR, "Database identifiers mismatch: we connected to DB id %llu, but in \"%s\" we found id %llu",
+				(long long)source_conn_id, dest_pgdata, (long long)dest_id);
 		}
 	}
 
@@ -439,7 +439,7 @@ catchup_thread_runner(void *arg)
 
 		if (file->write_size == BYTES_INVALID)
 		{
-			elog(LOG, "Skipping the unchanged file: \"%s\", read %li bytes", from_fullpath, file->read_size);
+			elog(LOG, "Skipping the unchanged file: \"%s\", read %zu bytes", from_fullpath, file->read_size);
 			continue;
 		}
